+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Bug fixes:
 * Fix `FFI::Pointer.new(FFI::Pointer)` to keep the original Pointer alive like CRuby (#4347, @eregon).
 * Check that `Symbol`s are always created from a non-broken `String` (@eregon).
 * Fix converting the same mutable `String` to native memory concurrently (@eregon).
+* Fix keeping `Float` and big `Integer` elements of an `Array` converted to native storage by `RARRAY_PTR()` alive (@eregon).
 
 Compatibility: